- Creato Sito TerraDellaVisione - Corretto Duplicazione Sito (es: da 1 a 14)

+++ inizia_riso.app
@@ -2,9 +2,21 @@
 
 site=riso.app
 
-rmdir __newfreeplanet.app
-rmdir __popolodelnuovomondo.app
-rmdir __riso.app
+if [ -d "__freeplanet.app" ]; then
+  rmdir __freeplanet.app
+fi
+
+if [ -d "__riso.app" ]; then
+  rmdir __riso.app
+fi
+
+if [ -d "__popolodelnuovomondo.app" ]; then
+  rmdir __popolodelnuovomondo.app
+fi
+
+if [ -d "__terradellavisione.app" ]; then
+  rmdir __terradellavisione.app
+fi
 
 
 source __inizia.sh
